+++ src/main/java/ru/ifmo/cet/javabasics/BottleSong.java
@@ -1,43 +1,73 @@
 package ru.ifmo.cet.javabasics;
 
-/**
- * Нужно реализовать констурктор и метод, возвращающий слова песни про бутылки на стене.
- * <p>
- * Слова следующие:
- * <p>
- * 99 bottles of beer on the wall, 99 bottles of beer
- * Take one down, pass it around, 98 bottles of beer
- * 98 bottles of beer on the wall, 98 bottles of beer
- * Take one down, pass it around, 97 bottles of beer
- * 97 bottles of beer on the wall, 97 bottles of beer
- * Take one down, pass it around, 96 bottles of beer
- * 96 bottles of beer on the wall, 96 bottles of beer
- * Take one down, pass it around, 95 bottles of beer
- * 95 bottles of beer on the wall, 95 bottles of beer
- * ...
- * <p>
- * 3 bottles of beer on the wall, 3 bottles of beer
- * Take one down, pass it around, 2 bottles of beer
- * 2 bottles of beer on the wall, 2 bottles of beer
- * Take one down, pass it around, 1 bottles of beer
- * 1 bottle of beer on the wall, 1 bottle of beer
- * Take one down, pass it around, no more bottles of beer on the wall
- * No more bottles of beer on the wall, no more bottles of beer
- * Go to the store and buy some more, 99 bottles of beer on the wall
- * <p>
- * Дело усложняется тем, что текст песни переменный:
- * За раз может быть взято несколько бутылок.
- * Значение передается в качестве параметра конструктора
- * Нужно ограничить возможность взятия бутылок натуральным число не более 99 бутылок за раз.
- */
-public class BottleSong {
+public class BottleSong
+{
+    private static final String[]
+            Nums = {"", "one", "two", "three", "four", "five", "six", "seven", "eight", "nine",
+            "ten", "eleven", "twelve", "thirteen", "fourteen", "fifteen","sixteen", "seventeen", "eighteen", "nineteen",
+            "twenty", "thirty", "forty", "fifty", "sixty", "seventy", "eighty", "ninety"};
 
-    public BottleSong(int bottleTakenAtOnce) {
-        //TODO
+    private int bottleTakenAtOnce = 0;
+	
+    public BottleSong( int bottleTakenAtOnce )
+    {
+        this.bottleTakenAtOnce = bottleTakenAtOnce;
     }
 
-    public String getBottleSongLyrics() {
-        //TODO
-        throw new UnsupportedOperationException();
+    public String getBottleSongLyrics()
+    {
+        if (bottleTakenAtOnce < 1 || bottleTakenAtOnce > 99)
+            throw new IllegalArgumentException();
+
+        int bottlesLeft = 99;
+        String TheSong = "";
+
+        while (bottlesLeft >= bottleTakenAtOnce) {
+            TheSong += sBottlesLeft(bottlesLeft) + " of beer on the wall, " + sBottlesLeft(bottlesLeft) + " of beer.\nTake "
+                    + sBottlesTake(bottleTakenAtOnce) + " down and pass around, ";
+            bottlesLeft -= bottleTakenAtOnce;
+            TheSong += sBottlesLeft(bottlesLeft) + " of beer on the wall.\n";
+        }
+
+        if (bottlesLeft > 0)
+            TheSong += sBottlesLeft(bottlesLeft) + " of beer on the wall, " + sBottlesLeft(bottlesLeft) + " of beer.\nTake "
+                    + sBottlesTake(bottlesLeft) + " down and pass around, " + sBottlesLeft(0) + " of beer on the wall.\n";
+
+        TheSong += "No more bottles of beer on the wall, no more bottles of beer.\nGo to the store and buy some more, 99 bottles of beer on the wall.\n";
+
+        return TheSong;
+    }
+
+    private String sBottlesLeft( int amount ) {
+        String res = "";
+        if (amount > 0)
+            res += (new Integer(amount)).toString();
+        else
+            res += "no more";
+
+        res += " bottle";
+
+        if (amount != 1)
+            res += "s";
+
+        return res;
+    }
+
+    private String sBottlesTake( int amount )
+    {
+        if (amount < 20)
+            return sSafeNumsObtain(amount);
+
+        String res = sSafeNumsObtain(amount / 10 + 18);
+        if (amount % 10 != 0)
+            res += " " + sSafeNumsObtain(amount % 10);
+
+         return res;
+    }
+
+    private String sSafeNumsObtain( int index ) {
+        if (index >= 0 && index < Nums.length)
+            return Nums[index];
+        return "";
     }
 }
